+++ src/Definition/Definition.php
@@ -8,7 +8,10 @@ use ArgumentCountError;
 use League\Container\Argument\ArgumentInterface;
 use League\Container\Argument\ArgumentResolverInterface;
 use League\Container\Argument\ArgumentResolverTrait;
+use League\Container\Argument\LiteralArgument;
 use League\Container\Argument\LiteralArgumentInterface;
+use League\Container\Argument\ResolvableArgument;
+use League\Container\Argument\ResolvableArgumentInterface;
 use League\Container\ContainerAwareTrait;
 use League\Container\Exception\ContainerException;
 use Override;
@@ -17,6 +20,7 @@ use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 use ReflectionClass;
 use ReflectionException;
+use ReflectionNamedType;
 
 class Definition implements ArgumentResolverInterface, DefinitionInterface
 {
@@ -24,6 +28,9 @@ class Definition implements ArgumentResolverInterface, DefinitionInterface
     use ContainerAwareTrait;
 
     protected mixed $resolved = null;
+
+    /** @var array<string, string|object> */
+    protected array $contextualArguments = [];
 
     /**
      * @param array<int, mixed> $arguments
@@ -160,6 +167,20 @@ class Definition implements ArgumentResolverInterface, DefinitionInterface
         return $this->arguments;
     }
 
+    #[Override]
+    public function addContextualArgument(string $abstract, string|object $concrete): DefinitionInterface
+    {
+        $this->contextualArguments[Definition::normaliseAlias($abstract)] = $concrete;
+        return $this;
+    }
+
+    /** @return array<string, string|object> */
+    #[Override]
+    public function getContextualArguments(): array
+    {
+        return $this->contextualArguments;
+    }
+
     /** @return list<array{method: string, arguments: array<int, mixed>}> */
     #[Override]
     public function getMethodCalls(): array
@@ -262,7 +283,13 @@ class Definition implements ArgumentResolverInterface, DefinitionInterface
      */
     protected function resolveClass(string $concrete): object
     {
-        $resolved   = $this->resolveArguments($this->arguments);
+        $arguments = $this->arguments;
+
+        if ($arguments === [] && $this->contextualArguments !== []) {
+            $arguments = $this->buildContextualArguments($concrete);
+        }
+
+        $resolved   = $this->resolveArguments($arguments);
         $reflection = new ReflectionClass($concrete);
 
         try {
@@ -276,6 +303,61 @@ class Definition implements ArgumentResolverInterface, DefinitionInterface
                 $concrete,
             ), 0, $e);
         }
+    }
+
+    /**
+     * @param class-string $concrete
+     * @return list<ResolvableArgumentInterface|LiteralArgument>
+     * @throws ContainerException
+     * @throws ReflectionException
+     */
+    protected function buildContextualArguments(string $concrete): array
+    {
+        $reflection  = new ReflectionClass($concrete);
+        $constructor = $reflection->getConstructor();
+
+        if ($constructor === null) {
+            return [];
+        }
+
+        $arguments = [];
+
+        foreach ($constructor->getParameters() as $param) {
+            $type = $param->getType();
+
+            if ($type instanceof ReflectionNamedType && !$type->isBuiltin()) {
+                $typeName = Definition::normaliseAlias($type->getName());
+
+                if (isset($this->contextualArguments[$typeName])) {
+                    $contextualConcrete = $this->contextualArguments[$typeName];
+
+                    if (is_object($contextualConcrete)) {
+                        $arguments[] = new LiteralArgument($contextualConcrete, null);
+                    } else {
+                        $arguments[] = new ResolvableArgument($contextualConcrete);
+                    }
+
+                    continue;
+                }
+
+                $arguments[] = new ResolvableArgument($typeName);
+                continue;
+            }
+
+            if ($param->isDefaultValueAvailable()) {
+                $arguments[] = new LiteralArgument($param->getDefaultValue(), null);
+                continue;
+            }
+
+            throw new ContainerException(sprintf(
+                'Cannot resolve parameter ($%s) of type (%s) for contextual binding on (%s): no contextual argument and no default value',
+                $param->getName(),
+                $type instanceof ReflectionNamedType ? $type->getName() : 'unknown',
+                $this->getAlias(),
+            ));
+        }
+
+        return $arguments;
     }
 
     /**