+++ src/ServiceProvider/ServiceProviderAggregate.php
@@ -7,6 +7,7 @@ namespace League\Container\ServiceProvider;
 use Generator;
 use League\Container\ContainerAwareTrait;
 use League\Container\Exception\ContainerException;
+use LogicException;
 use Override;
 
 class ServiceProviderAggregate implements ServiceProviderAggregateInterface
@@ -48,11 +49,40 @@ class ServiceProviderAggregate implements ServiceProviderAggregateInterface
         return false;
     }
 
+    #[Override]
+    public function providerClassFor(string $id): string
+    {
+        foreach ($this as $provider) {
+            if ($provider->provides($id)) {
+                return $provider::class;
+            }
+        }
+
+        throw new LogicException(sprintf('No provider claims to provide (%s)', $id));
+    }
+
     /** @return Generator<int, ServiceProviderInterface> */
     #[Override]
     public function getIterator(): Generator
     {
         yield from $this->providers;
+    }
+
+    /** @return list<string> */
+    #[Override]
+    public function getProvidedIds(): array
+    {
+        $ids = [];
+
+        foreach ($this as $provider) {
+            foreach ($provider->getProvidedIds() as $id) {
+                if (!in_array($id, $ids, strict: true)) {
+                    $ids[] = $id;
+                }
+            }
+        }
+
+        return $ids;
     }
 
     #[Override]