+++ src/ServiceProvider/ServiceProviderAggregateInterface.php
@@ -12,6 +12,11 @@ interface ServiceProviderAggregateInterface extends ContainerAwareInterface, Ite
 {
     public function add(ServiceProviderInterface $provider): ServiceProviderAggregateInterface;
     public function provides(string $id): bool;
+    public function providerClassFor(string $id): string;
+
+    /** @return list<string> */
+    public function getProvidedIds(): array;
+
     public function register(string $service): void;
     public function registerAll(): void;
 }