+++ tests/Asset/SharedService.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace League\Container\Test\Asset;
+
+use League\Container\Attribute\Shared;
+
+#[Shared]
+class SharedService {}